Adds attrs = {} to Atm.withdraw and tests person-atm interaction
atm.withdraw used to take regular parameters, but I changed it to take *args to a hash so that I could evaluate entries without necessarily needing things to be perfect or in the right order. This required a number of name changes throughout Atm but everything running fine now.
Also created an integration file to test that person and atm were talking to each other properly.